Moved Resource to core
